move Region definition and add Region references to KNode
