+++ packages/klighd-interactive/src/constraint-classes.ts
@@ -15,7 +15,7 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 
-import { moveFeature, Point, RectangularNode, SEdge, selectFeature, SParentElement } from 'sprotty/lib';
+import { Bounds, moveFeature, Point, RectangularNode, SEdge, selectFeature, SParentElement } from 'sprotty/lib';
 
 /**
  * This is the superclass of all elements of a graph such as nodes, edges, ports,
@@ -62,7 +62,19 @@ export class KNode extends RectangularNode implements KGraphElement {
     shadowX: number
     shadowY: number
 
-    detailReference: DetailReference
+
+    /** 
+     * The Region if any that contains this KNode 
+     * If this is undefined it has yet to be created/checked
+     * If this is null the Region is a Region root
+     */
+    containingRegion?: Region | null
+
+    /** The Region if any that is constructed by the KNode 
+     *  If this is undefined it has yet to be created/checked
+     *  If this is null this KNode does not construct a region
+     */
+    providingRegion?: Region | null
 }
 
 export enum Direction {
@@ -87,6 +99,45 @@ export enum DetailLevel {
  */
 export interface DetailReference {
     detailLevel: DetailLevel
+}
+
+/**
+ * Combines KNodes into regions. These correspond to child areas. A region can correspond to 
+ * a region or a super state in the model. Also manages the boundaries, title candidates, 
+ * tree structure of the model and application of detail level of its KNodes.
+ */
+export class Region {
+    /** All KNodes specifically in the region. */
+    elements: KNode[]
+    /** The rectangle of the child area in which the region lies. */
+    boundingRectangle: KNode
+    /** Gained using browser position and rescaling and are therefore not perfect. */
+    absoluteBounds: Bounds
+    /** the regions current detail level that is used by all children */
+    detail: DetailLevel
+    /** The immediate parent region of this region. */
+    parent?: Region
+    /** All immediate child regions of this region */
+    children: Region[]
+    /** Contains the height of the title of the region, if there is one. */
+    regionTitleHeight?: number
+
+    /** Constructor initializes element array for region. */
+    constructor(boundingRectangle: KNode) {
+        this.boundingRectangle = boundingRectangle
+        this.elements = []
+        this.children = []
+        this.detail = DetailLevel.FullDetails
+        boundingRectangle.providingRegion = this
+    }
+
+    /** 
+     * Applies the detail level to all elements of a region.
+     * @param level the detail leveel to apply
+     */
+    setDetailLevel(level: DetailLevel): void {
+        this.detail = level
+    }
 }
 
 /**
